Fix MSVC stack overflow during exception unwinding with CallTraceStack
MSVC exception handling uses ~17KB of stack space per level when
unwinding through catch-rethrow patterns. With deep recursion (1000+
levels), this causes stack overflow during the unwind phase, even
though the recursion itself only uses ~725 bytes per level.

Solution:
- Add CallTraceStack: A thread-local trace collection using RAII guards
  that uses std::uncaught_exceptions() to preserve trace during exception
  propagation without catching/rethrowing at each level
- Replace all catch-modify-throw patterns in recursive evaluation code
  (Expression.cc, UserModule.cc, ModuleInstantiation.cc, ScopeContext.cc,
  Value.cc) with CallTraceStack::Guard
- Update exceptions.h with printCallTrace() to format trace output
- Use GetCurrentThreadStackLimits() Windows API for accurate stack
  boundary detection in StackCheck
- Add OPENSCAD_MSVC_RECURSION_LIMIT CMake cache variable (default 1000)
- Add vectored exception handler as safety net for stack overflow
- Add msvc-recursion-limit test

Fixes stack overflow crashes when evaluating deeply recursive OpenSCAD
code on Windows/MSVC builds.

+++ src/platform/PlatformUtils-posix.cc
@@ -122,7 +122,7 @@ std::string PlatformUtils::userConfigPath()
   return "";
 }
 
-size_t PlatformUtils::stackLimit()
+unsigned long PlatformUtils::stackLimit()
 {
 #ifndef __EMSCRIPTEN__
   struct rlimit limit;
@@ -133,13 +133,13 @@ size_t PlatformUtils::stackLimit()
       return STACK_LIMIT_DEFAULT;
     }
     if (limit.rlim_cur > STACK_BUFFER_SIZE) {
-      return static_cast<size_t>(limit.rlim_cur - STACK_BUFFER_SIZE);
+      return static_cast<unsigned long>(limit.rlim_cur - STACK_BUFFER_SIZE);
     }
     if (limit.rlim_max == RLIM_INFINITY) {
       return STACK_LIMIT_DEFAULT;
     }
     if (limit.rlim_max > STACK_BUFFER_SIZE) {
-      return static_cast<size_t>(limit.rlim_max - STACK_BUFFER_SIZE);
+      return static_cast<unsigned long>(limit.rlim_max - STACK_BUFFER_SIZE);
     }
   }
 #endif  // __EMSCRIPTEN__
